The objective of this fork is to allow the sickle project to use multi-threading

+++ src/kseq.h
@@ -115,12 +115,12 @@ typedef struct __kstring_t {
                for (i = ks->begin; i < ks->end; ++i)                    \
                  if (isspace(ks->buf[i]) && ks->buf[i] != ' ') break;   \
              } else i = 0; /* never come to here! */                    \
-               if (str->m - str->l < i - ks->begin + 1) {               \
-               str->m = str->l + (i - ks->begin) + 1;                   \
-               kroundup32(str->m);                                      \
-               str->s = (char*)realloc(str->s, str->m);                 \
-             }                                                          \
-               memcpy(str->s + str->l, ks->buf + ks->begin, i - ks->begin); \
+               if ((str->m - str->l) < (size_t)(i - ks->begin + 1)) {               \
+                str->m = str->l + (i - ks->begin) + 1;                  \
+                kroundup32(str->m);                                     \
+                str->s = (char*)realloc(str->s, str->m);                \
+               }                                                        \
+               memcpy(str->s + str->l, ks->buf + ks->begin, i - ks->begin);\
                str->l = str->l + (i - ks->begin);                       \
                ks->begin = i + 1;                                       \
                if (i < ks->end) {                                       \
